refactor: reorganize browser use folder to include browser related tools
