+++ packages/shortest/scripts/test-browser.ts
@@ -1,5 +1,5 @@
-import { BrowserManager } from '../src/core/browser-manager';
-import { BrowserTool } from '../src/browser-use/browser';
+import { BrowserManager } from '../src/browser/browser-manager';
+import { BrowserTool } from '../src/browser/browser';
 import { initialize } from '../src/index';
 import pc from 'picocolors';
