+++ packages/shortest/scripts/test-github.ts
@@ -1,6 +1,6 @@
-import { BrowserManager } from '../src/core/browser-manager';
-import { BrowserTool } from '../src/browser-use/browser';
-import { GitHubTool } from '../src/tools/github';
+import { BrowserManager } from '../src/browser/browser-manager';
+import { BrowserTool } from '../src/browser/browser';
+import { GitHubTool } from '../src/browser/integrations/github';
 import { initialize } from '../src/index';
 import pc from 'picocolors';
