+++ packages/shortest/scripts/test-loop.ts
@@ -1,5 +1,5 @@
-import { BrowserManager } from '../src/core/browser-manager';
-import { BrowserTool } from '../src/browser-use/browser';
+import { BrowserManager } from '../src/browser/browser-manager';
+import { BrowserTool } from '../src/browser/browser';
 import { defaultConfig, initialize } from '../src/index';
 import Anthropic from '@anthropic-ai/sdk';
 import { SYSTEM_PROMPT } from '../src/ai/system-prompt';
