+++ packages/shortest/src/browser/base.ts
